requirements.txt
aiofiles==24.1.0
aiogram==3.21.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
annotated-types==0.7.0
asgiref==3.9.1
attrs==25.3.0
Brotli==1.1.0
certifi==2025.7.9
cffi==1.17.1
charset-normalizer==3.4.3
cssselect2==0.8.0
Django==5.2.4
django-cors-headers==4.7.0
django-jazzmin==3.0.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.11
fonttools==4.58.5
frozenlist==1.7.0
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
magic-filter==1.0.12
multidict==6.6.3
nest-asyncio==1.6.0
packaging==25.0
pillow==11.3.0
propcache==0.3.2
psycopg2==2.9.10
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.11.7
pydantic_core==2.33.2
pydyf==0.11.0
PyJWT==2.9.0
pyphen==0.17.2
python-authorizer @ git+https://github.com/mancho-devs/python-authorizer.git@31c34e05cc476b2237b60a35f7be52b4479c1139
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.3
reportlab==4.4.3
requests==2.32.5
sqlparse==0.5.3
tinycss2==1.4.0
tinyhtml5==2.0.0
typing-inspection==0.4.1
typing_extensions==4.14.1
uritemplate==4.2.0
urllib3==2.5.0
weasyprint==65.1
webencodings==0.5.1
yarl==1.20.1
zopfli==0.2.3.post1
